create access_control_helper module, various cleanup
