+++ .github/workflows/pipeline_01_staging.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   staging:
     name: Stag
-    # if: ${{ github.ref_type == 'branch' && github.ref == 'refs/heads/master'}}
+    # if: ${{ github.ref_type == 'branch' && github.ref == 'refs/heads/main'}}
     uses: ./.github/workflows/pipeline_02_structure.yml
     with:
       tags: type=sha
