+++ app/views/programs/show/_topics_matrix.html.erb
@@ -24,7 +24,7 @@ Matrix
           </td>
       <% else %>
         <td >
-          <%= link_to "+", new_topic_description_path(course_id: course.id, topic_id: topic.id, back_to: course.id)  %>
+          <%= link_to_new "+", TopicDescription, course_id: course.id, topic_id: topic.id, back_to: course.id  %>
         </td>
       <% end %>
